requirements.txt
fastapi
uvicorn
python-dotenv
langchain
chromadb
langchain-community
langchain-core
langchain-openai